backend/reports/permissions.py
from rest_framework import permissions

class IsOwnerOrManager(permissions.BasePermission):
    def has_permission(self, request, view):
        return request.user.is_authenticated and request.user.role in ['owner', 'manager']